+++ malis/malis_cpp.cpp
@@ -18,7 +18,7 @@ class AffinityGraphCompare{
 		AffinityGraphCompare(const T * EdgeWeightArray){
 			mEdgeWeightArray = EdgeWeightArray;
 		}
-		bool operator() (const int& ind1, const int& ind2) const {
+		bool operator() (const uint64_t& ind1, const uint64_t& ind2) const {
 			return (mEdgeWeightArray[ind1] > mEdgeWeightArray[ind2]);
 		}
 };
@@ -29,8 +29,8 @@ class AffinityGraphCompare{
  * Author: Srini Turaga (sturaga@mit.edu)
  * All rights reserved
  */
-void malis_loss_weights_cpp(const int nVert, const uint64_t* seg,
-               const int nEdge, const uint64_t* node1, const uint64_t* node2, const float* edgeWeight,
+void malis_loss_weights_cpp(const uint64_t nVert, const uint64_t* seg,
+               const uint64_t nEdge, const uint64_t* node1, const uint64_t* node2, const float* edgeWeight,
                const int pos,
                uint64_t* nPairPerEdge){
 
@@ -40,7 +40,7 @@ void malis_loss_weights_cpp(const int nVert, const uint64_t* seg,
     vector<uint64_t> rank(nVert);
     vector<uint64_t> parent(nVert);
     boost::disjoint_sets<uint64_t*, uint64_t*> dsets(&rank[0],&parent[0]);
-    for (int i=0; i<nVert; ++i){
+    for (uint64_t i=0; i<nVert; ++i){
         dsets.make_set(i);
         if (0!=seg[i]) {
             overlap[i].insert(pair<uint64_t,uint64_t>(seg[i],1));
@@ -48,9 +48,9 @@ void malis_loss_weights_cpp(const int nVert, const uint64_t* seg,
     }
 
     /* Sort all the edges in increasing order of weight */
-    std::vector< int > pqueue( nEdge );
-    int j = 0;
-    for ( int i = 0; i < nEdge; i++ ){
+    std::vector< uint64_t > pqueue( nEdge );
+    uint64_t j = 0;
+    for ( uint64_t i = 0; i < nEdge; i++ ){
         if ((node1[i]>=0) && (node1[i]<nVert) && (node2[i]>=0) && (node2[i]<nVert))
 	        pqueue[ j++ ] = i;
     }
@@ -60,13 +60,13 @@ void malis_loss_weights_cpp(const int nVert, const uint64_t* seg,
 
 
     /* Start MST */
-    int e;
+    uint64_t e;
     uint64_t set1, set2;
     uint64_t nPair = 0;
     map<uint64_t,uint64_t>::iterator it1, it2;
 
     /* Start Kruskal's */
-    for (unsigned int i = 0; i < pqueue.size(); ++i ) {
+    for (uint64_t i = 0; i < pqueue.size(); ++i ) {
         e = pqueue[i];
 
         set1 = dsets.find_set(node1[e]);
@@ -111,31 +111,31 @@ void malis_loss_weights_cpp(const int nVert, const uint64_t* seg,
 }
 
 
-void connected_components_cpp(const int nVert,
-               const int nEdge, const uint64_t* node1, const uint64_t* node2, const int* edgeWeight,
+void connected_components_cpp(const uint64_t nVert,
+               const uint64_t nEdge, const uint64_t* node1, const uint64_t* node2, const int* edgeWeight,
                uint64_t* seg){
 
     /* Make disjoint sets */
     vector<uint64_t> rank(nVert);
     vector<uint64_t> parent(nVert);
     boost::disjoint_sets<uint64_t*, uint64_t*> dsets(&rank[0],&parent[0]);
-    for (int i=0; i<nVert; ++i)
+    for (uint64_t i=0; i<nVert; ++i)
         dsets.make_set(i);
 
     /* union */
-    for (int i = 0; i < nEdge; ++i )
+    for (uint64_t i = 0; i < nEdge; ++i )
          // check bounds to make sure the nodes are valid
         if ((edgeWeight[i]!=0) && (node1[i]>=0) && (node1[i]<nVert) && (node2[i]>=0) && (node2[i]<nVert))
             dsets.union_set(node1[i],node2[i]);
 
     /* find */
-    for (int i = 0; i < nVert; ++i)
+    for (uint64_t i = 0; i < nVert; ++i)
         seg[i] = dsets.find_set(i);
 }
 
 
-void marker_watershed_cpp(const int nVert, const uint64_t* marker,
-               const int nEdge, const uint64_t* node1, const uint64_t* node2, const float* edgeWeight,
+void marker_watershed_cpp(const uint64_t nVert, const uint64_t* marker,
+               const uint64_t nEdge, const uint64_t* node1, const uint64_t* node2, const float* edgeWeight,
                uint64_t* seg){
 
     /* Make disjoint sets */
@@ -159,9 +159,9 @@ void marker_watershed_cpp(const int nVert, const uint64_t* marker,
             dsets.union_set(components[seg[i]],i);
 
     /* Sort all the edges in decreasing order of weight */
-    std::vector<int> pqueue( nEdge );
-    int j = 0;
-    for (int i = 0; i < nEdge; ++i)
+    std::vector<uint64_t> pqueue( nEdge );
+    uint64_t j = 0;
+    for (uint64_t i = 0; i < nEdge; ++i)
         if ((edgeWeight[i]!=0) &&
             (node1[i]>=0) && (node1[i]<nVert) &&
             (node2[i]>=0) && (node2[i]<nVert) &&
@@ -172,9 +172,9 @@ void marker_watershed_cpp(const int nVert, const uint64_t* marker,
     sort( pqueue.begin(), pqueue.end(), AffinityGraphCompare<float>( edgeWeight ) );
 
     /* Start MST */
-	int e;
-    int set1, set2, label_of_set1, label_of_set2;
-    for (unsigned int i = 0; i < pqueue.size(); ++i ) {
+	uint64_t e;
+    uint64_t set1, set2, label_of_set1, label_of_set2;
+    for (uint64_t i = 0; i < pqueue.size(); ++i ) {
 		e = pqueue[i];
         set1=dsets.find_set(node1[e]);
         set2=dsets.find_set(node2[e]);
@@ -194,7 +194,7 @@ void marker_watershed_cpp(const int nVert, const uint64_t* marker,
     }
 
     // write out the final coloring
-    for (int i=0; i<nVert; i++)
+    for (uint64_t i=0; i<nVert; i++)
         seg[i] = seg[dsets.find_set(i)];
 
 }
